Moved functions into actions, out of tree
